+++ ios/ViroReact/VRTVideoSurfaceManager.mm
@@ -63,7 +63,7 @@ RCT_EXPORT_VIEW_PROPERTY(ignoreEventHandling, BOOL)
     return [[VRTVideoSurface alloc] initWithBridge:self.bridge];
 }
 
-RCT_EXPORT_METHOD(seekToTime:(nonnull NSNumber *)reactTag time:(NSInteger)time)
+RCT_EXPORT_METHOD(seekToTime:(nonnull NSNumber *)reactTag time:(float)time)
 {
     [self.bridge.uiManager addUIBlock:^(__unused RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         VRTView *view = (VRTView *)viewRegistry[reactTag];
